Refactor: daemons: Rearrange existing IPC code in scheduler daemon.
This brings it closer in line with how the controld and pacemakerd
daemons are arranged.
